Move TimeOfYearSelector to its own directory
