+++ src/components/TimeOfYearSelector/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "TimeOfYearSelector",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./TimeOfYearSelector.js"
+}
